+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: gakarbou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/22 21:07:29 by gakarbou          #+#    #+#             */
-/*   Updated: 2025/03/05 10:18:50 by lroussel         ###   ########.fr       */
+/*   Updated: 2025/03/05 17:16:07 by gakarbou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
