+++ src/readline/lines.c
@@ -6,7 +6,7 @@
 /*   By: lroussel <lroussel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/10 11:22:28 by lroussel          #+#    #+#             */
-/*   Updated: 2025/03/10 11:24:54 by lroussel         ###   ########.fr       */
+/*   Updated: 2025/03/11 12:20:01 by lroussel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ static int	count_low_newlines_prompt(t_readline *data, int *j, t_vector2 size)
 		{
 			count += *j / size.x - (data->prompt[i] == '\n'
 					* (*j % size.x == 1) * (data->prompt[i - 1] != '\n'));
-			j = 0;
+			*j = 0;
 		}
 		i++;
 	}
@@ -44,7 +44,7 @@ int	count_low_newlines(t_readline *data, t_char *to)
 	size = get_terminal_size(data, 0);
 	count = count_low_newlines_prompt(data, &j, size);
 	if ((to && !data->actual) || (data->first && !data->actual))
-		return (0);
+		return (count);
 	c = data->first;
 	while (c && (to == NULL || c != to->next))
 	{
@@ -72,7 +72,7 @@ int	count_hard_newlines(t_readline data, t_char *to)
 		count += (data.prompt[i++] == '\n');
 	c = data.first;
 	if ((to && !data.actual) || (data.first && !data.actual))
-		return (0);
+		return (count);
 	while (c && (to == NULL || c != to->next))
 	{
 		count += c->c == '\n';
